Use common MetaOutput component
