+++ src/components/questionnaire-builder/MetaOutput/styles.scss
@@ -1,0 +1,17 @@
+.meta-output {
+    display: inline-flex;
+    align-items: baseline;
+    padding: var(--spacing-extra-small) 0;
+    color: var(--color-text-label);
+
+    .name {
+        padding: 0 var(--spacing-small);
+    }
+
+    &:not(:last-child) {
+        &:after {
+            content: '•';
+        }
+    }
+}
+
